pyproject.toml
[tool.poetry]
name = "cardiospike"
version = "0.1.0"
description = ""
authors = ["gleb <gleberof@gmail.com>", "Daniil Gafni <danielgafni16@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.6"
sklearn = "^0.0"
pandas = "^1.2.4"
numpy = "^1.20.3"
matplotlib = "^3.4.2"
scipy = "^1.6.3"
torch = "^1.8.1"
lightgbm = "^3.2.1"
catboost = "^0.26"
xgboost = "^1.4.2"
tqdm = "^4.61.0"
transformers = "^4.6.1"
pytorch-lightning = "*"
Cython = "^0.29.23"
hydra-core = "^1.1.0"
optuna = "^2.8.0"
plotly = "^4.14.3"
uvicorn = "^0.14.0"
fastapi = "^0.65.2"
python-dotenv = "^0.17.1"
Flask = "^2.0.1"
dash = "^1.20.0"
gunicorn = "^20.1.0"
importlib = "^1.0.4"
types-requests = "^0.1.11"
s3fs = "0.4"
boto3 = "^1.16.35"
mysql-connector = "^2.2.9"
mysqlclient = "^2.0.3"
PyMySQL = "^1.0.2"
flask-swagger-ui = "^3.36.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
mypy = "^0.902"
flake8 = "^3.9.2"
isort = "^5.8.0"
autoflake = "^1.4"
pytest = "^6.2.4"
pre-commit = "^2.13.0"
black = "^21.6b0"
poethepoet = "^0.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | build
  | dist
  | configs
  )/
'''

[tool.isort]
profile = "black"
line_length = 120
src_paths = ["cardiospike", "tests"]

[tool.pytest.ini_options]
testpaths = [
    "tests"
]

[tool.poe.tasks]
force-cuda11 = "python -m pip install torch==1.8.1+cu111 torchvision==0.9.1+cu111 -f https://download.pytorch.org/whl/torch_stable.html"
